Add documentation, remove linter check
