+++ crates/crust-render/src/camera.rs
@@ -1,17 +1,39 @@
 use crate::ray::Ray;
 use utils::{Point3, Vec3};
 
+/// The `Camera` struct represents a virtual camera in the ray tracing system.
+/// It is responsible for generating rays that simulate the perspective view of a scene.
 pub struct Camera {
+    /// The origin of the camera (position in 3D space).
     origin: Point3,
+    /// The lower-left corner of the viewport.
     lower_left_corner: Point3,
+    /// The horizontal vector of the viewport.
     horizontal: Vec3,
+    /// The vertical vector of the viewport.
     vertical: Vec3,
+    /// The camera's local horizontal axis.
     u: Vec3,
+    /// The camera's local vertical axis.
     v: Vec3,
+    /// The radius of the camera's lens (used for depth of field).
     lens_radius: f32,
 }
 
 impl Camera {
+    /// Creates a new `Camera` with the specified parameters.
+    ///
+    /// # Parameters
+    /// - `lookfrom`: The position of the camera in 3D space.
+    /// - `lookat`: The point in 3D space the camera is looking at.
+    /// - `vup`: The "up" direction vector for the camera.
+    /// - `vfov`: The vertical field-of-view in degrees.
+    /// - `aspect_ratio`: The aspect ratio of the viewport (width/height).
+    /// - `aperture`: The aperture size of the camera (controls depth of field).
+    /// - `focus_dist`: The distance to the focus plane.
+    ///
+    /// # Returns
+    /// - A new instance of `Camera`.
     pub fn new(
         lookfrom: Point3,
         lookat: Point3,
@@ -47,6 +69,14 @@ impl Camera {
         }
     }
 
+    /// Generates a ray originating from the camera through the viewport.
+    ///
+    /// # Parameters
+    /// - `s`: The horizontal coordinate on the viewport (normalized to [0, 1]).
+    /// - `t`: The vertical coordinate on the viewport (normalized to [0, 1]).
+    ///
+    /// # Returns
+    /// - A `Ray` that starts at the camera and passes through the specified point on the viewport.
     pub fn get_ray(&self, s: f32, t: f32) -> Ray {
         let rd = self.lens_radius * utils::random_in_unit_disk();
         let offset = self.u * rd.x() + self.v * rd.y();
