+++ crates/crust-render/src/lib.rs
@@ -1,21 +1,22 @@
-pub mod buffer;
-pub mod camera;
-pub mod convert;
-pub mod hittable;
-pub mod hittable_list;
-pub mod light;
-pub mod material;
-pub mod primitives;
-pub mod ray;
+mod buffer;
+mod camera;
+mod convert;
+mod hittable;
+mod hittable_list;
+mod light;
+mod material;
+mod primitives;
+mod ray;
 mod sampler;
-pub mod tracer;
-pub mod world;
+mod tracer;
+mod world;
 
-pub use buffer::Buffer;
 pub use camera::Camera;
+pub use convert::convert;
 pub use hittable_list::HittableList;
 pub use light::{Light, LightList};
 pub use material::*;
 pub use ray::Ray;
 pub use sampler::generate_cmj_2d;
-pub use world::random_scene;
+pub use tracer::{RenderSettings, Renderer};
+pub use world::{random_scene, simple_scene};
